Add files via upload
Files needed for analysis. 

The goal is to determine the effectiveness of the ad campaigns in driving clicks and sales. The company also needs to understand which product categories are most popular among different customer segments, which campaigns are performing best by gender, and which age groups are most likely to engage with marketing content.